requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.6.0
scikit-learn>=1.3.0
numpy>=1.23.0
scipy>=1.10.0
PyPDF2>=3.0.0
docx2txt>=0.8
unidecode>=1.3.6
python-dotenv>=1.0.0

sentence-transformers>=2.2.2 
torch>=2.0.0 
Jinja2>=3.1.0
Jinja2>=3.1.0
Jinja2>=3.1.0
reportlab>=4.0.0
